Updated the PayeEstimator to include an optional versionDir.
The versionDir param is required when we would like to upgrade an
existing js lib, the new js lib in this would replace a previous
version and by specifying the versionDir you would like to replace. If
this versionDir is not supplied then the default behaviour would be to
put the new js lib in a directory with the same name as the artefact
version.